% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_numbers}
\alias{add_numbers}
\title{add two numbers}
\usage{
add_numbers(input1, input2)
}
\arguments{
\item{input1}{int A number}

\item{input2}{int A number}
}
\value{
The sum of \code{input1} and \code{input2}.
}
\description{
adds things
}
\details{
This part can be quite long

and several paragraphs
}
\examples{
add_numbers(2, 3)

}
